refactor: Deduplicate a bunch of code in TCP client/server.
Also generally a bit of cleanup and better layering.
