+++ CMakeLists.txt
@@ -198,6 +198,8 @@ set(toxcore_SOURCES ${toxcore_SOURCES}
 set(toxcore_SOURCES ${toxcore_SOURCES}
   toxcore/TCP_client.c
   toxcore/TCP_client.h
+  toxcore/TCP_common.c
+  toxcore/TCP_common.h
   toxcore/TCP_connection.c
   toxcore/TCP_connection.h
   toxcore/TCP_server.c
