+++ auto_tests/BUILD.bazel
@@ -39,7 +39,10 @@ flaky_tests = {
         "//c-toxcore/toxcore",
         "//c-toxcore/toxcore:DHT_srcs",
         "//c-toxcore/toxcore:Messenger",
+        "//c-toxcore/toxcore:TCP_client",
+        "//c-toxcore/toxcore:TCP_common",
         "//c-toxcore/toxcore:TCP_connection",
+        "//c-toxcore/toxcore:TCP_server",
         "//c-toxcore/toxcore:ccompat",
         "//c-toxcore/toxcore:crypto_core",
         "//c-toxcore/toxcore:friend_connection",
