+++ other/bootstrap_daemon/docker/tox-bootstrapd.sha256
@@ -1,1 +1,1 @@
-7be4fcdc11f74036b97de9569eb06f82296f193db3e18e671171f3bc059f95a6  /usr/local/bin/tox-bootstrapd
+fb46c678adbe48e846286d9cb45b560e26f51cb7eccb99378c57e66c6c49732b  /usr/local/bin/tox-bootstrapd
