+++ toxcore/Messenger.h
@@ -10,6 +10,7 @@
 #ifndef C_TOXCORE_TOXCORE_MESSENGER_H
 #define C_TOXCORE_TOXCORE_MESSENGER_H
 
+#include "TCP_server.h"
 #include "friend_connection.h"
 #include "friend_requests.h"
 #include "logger.h"
