+++ toxcore/TCP_client.h
@@ -9,8 +9,8 @@
 #ifndef C_TOXCORE_TOXCORE_TCP_CLIENT_H
 #define C_TOXCORE_TOXCORE_TCP_CLIENT_H
 
-#include "TCP_server.h"
-#include "crypto_core.h"
+#include "mono_time.h"
+#include "network.h"
 
 #define TCP_CONNECTION_TIMEOUT 10
