+++ toxcore/net_crypto.c
@@ -14,6 +14,7 @@
 #include <stdlib.h>
 #include <string.h>
 
+#include "list.h"
 #include "mono_time.h"
 #include "util.h"
